Latest updated from Srini:
Hi Sapan
Finally, I got everything to work fine. The two main issues I realized
late were that 1) the wsdl has changed to include the request_hash,
and 2) the SOAPpy module needs to be installed for SOAP to work in the
SFA server. All good now.

Please check in the attached updated version of the RSpec manager.

Thanks
Srini.